requirements.txt
langchain-core
langchain-openai
langgraph
requests
python-dotenv
openai

#API
fastapi
uvicorn 
pydantic

#LLM Guard
llm-guard